pyproject.toml
[project]
name = "aws-account-audit"
version = "0.1.0"
description = "AWS hygiene & security audit tool producing JSON/MD/HTML reports"
requires-python = ">=3.10"
dependencies = [
  "boto3>=1.34",
  "PyYAML>=6.0",
  "Jinja2>=3.1",
  "python-dotenv>=1.0",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "moto>=5.0",
  "ruff>=0.6",
  "black>=24.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
